+++ Ppap/app/src/main/java/com/jeongg/ppap/presentation/component/PTabLayer.kt
@@ -11,8 +11,8 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
-import com.jeongg.ppap.data.notice.dto.SubscribeDTO
-import com.jeongg.ppap.presentation.theme.main_yellow
+import com.jeongg.ppap.data.dto.SubscribeDTO
+import com.jeongg.ppap.theme.main_yellow
 
 @Composable
 fun PTabLayer(
