+++ Ppap/app/src/main/java/com/jeongg/ppap/presentation/component/PTitle.kt
@@ -20,8 +20,8 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import com.jeongg.ppap.R
-import com.jeongg.ppap.presentation.theme.Dimens
-import com.jeongg.ppap.presentation.theme.gray3
+import com.jeongg.ppap.theme.Dimens
+import com.jeongg.ppap.theme.gray3
 
 @Composable
 fun PTitle(
